requirements.txt
pandas
scikit-learn
streamlit
matplotlib
seaborn
python-docx